LayoutTests/inspector/canvas/recording-2d-expected.txt

Test that CanvasManager is able to record actions made to 2D canvas contexts.


== Running test suite: Canvas.recording2D
-- Running test case: Canvas.recording2D.singleFrame
initialState:
  attributes:
    width: 2
    height: 2
    setTransform: [1,0,0,1,0,0]
    globalAlpha: 1
    globalCompositeOperation: 0
    lineWidth: 1
    lineCap: 1
    lineJoin: 2
    miterLimit: 10
    shadowOffsetX: 0
    shadowOffsetY: 0
    shadowBlur: 0
    shadowColor: 3
    setLineDash: [[]]
    lineDashOffset: 0
    font: 4
    textAlign: 5
    textBaseline: 6
    direction: 7
    strokeStyle: 8
    fillStyle: 8
    imageSmoothingEnabled: true
    imageSmoothingQuality: 9
    setPath: [10]
  parameters:
  content: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAAAtJREFUCB1jYEAHAAASAAGAFMrMAAAAAElFTkSuQmCC"
frames:
  0:
    0: arc(1, 2, 3, 4, 5, false)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:77:42 (13)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:77:28 (16)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
        4: performActions - inspector/canvas/recording-2d.html:394:25 (20)
        5: Global Code - [program code] (22)
        6: evaluateWithScopeExtension - [native code] (25)
        7: _evaluateOn (27)
        8: _evaluateAndWrap (29)
    1: arc(6, 7, 8, 9, 10, true)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:78:42 (30)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:78:28 (31)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
        4: performActions - inspector/canvas/recording-2d.html:394:25 (20)
        5: Global Code - [program code] (22)
        6: evaluateWithScopeExtension - [native code] (25)
        7: _evaluateOn (27)
        8: _evaluateAndWrap (29)
data:
  0: "source-over"
  1: "butt"
  2: "miter"
  3: "rgba(0, 0, 0, 0)"
  4: "10px sans-serif"
  5: "start"
  6: "alphabetic"
  7: "ltr"
  8: "#000000"
  9: "low"
  10: ""
  11: "arc"
  12: "inspector/canvas/recording-2d.html"
  13: [10,12,78,42]
  14: "ignoreException"
  15: [14,12,49,13]
  16: [10,12,78,28]
  17: "executeFrameFunction"
  18: [17,12,391,24]
  19: "performActions"
  20: [19,12,395,25]
  21: "global code"
  22: [21,10,1,15]
  23: "evaluateWithScopeExtension"
  24: "[native code]"
  25: [23,24,0,0]
  26: "_evaluateOn"
  27: [26,10,128,29]
  28: "_evaluateAndWrap"
  29: [28,10,122,108]
  30: [10,12,79,42]
  31: [10,12,79,28]

-- Running test case: Canvas.recording2D.multipleFrames
initialState:
  attributes:
    width: 2
    height: 2
    setTransform: [1,0,0,1,0,0]
    globalAlpha: 1
    globalCompositeOperation: 0
    lineWidth: 1
    lineCap: 1
    lineJoin: 2
    miterLimit: 10
    shadowOffsetX: 0
    shadowOffsetY: 0
    shadowBlur: 0
    shadowColor: 3
    setLineDash: [[]]
    lineDashOffset: 0
    font: 4
    textAlign: 5
    textBaseline: 6
    direction: 7
    strokeStyle: 8
    fillStyle: 8
    imageSmoothingEnabled: true
    imageSmoothingQuality: 9
    setPath: [10]
  parameters:
  content: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAAAtJREFUCB1jYEAHAAASAAGAFMrMAAAAAElFTkSuQmCC"
frames:
  0:
    0: arc(1, 2, 3, 4, 5, false)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:77:42 (13)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:77:28 (16)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
        4: performActions - inspector/canvas/recording-2d.html:394:25 (20)
        5: Global Code - [program code] (22)
        6: evaluateWithScopeExtension - [native code] (25)
        7: _evaluateOn (27)
        8: _evaluateAndWrap (29)
    1: arc(6, 7, 8, 9, 10, true)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:78:42 (30)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:78:28 (31)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
        4: performActions - inspector/canvas/recording-2d.html:394:25 (20)
        5: Global Code - [program code] (22)
        6: evaluateWithScopeExtension - [native code] (25)
        7: _evaluateOn (27)
        8: _evaluateAndWrap (29)
  1:
    0: arcTo(1, 2, 3, 4, 5)
      swizzleTypes: [Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:81:44 (33)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:81:28 (34)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  2:
    0: beginPath()
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:84:26 (36)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  3:
    0: bezierCurveTo(1, 2, 3, 4, 5, 6)
      swizzleTypes: [Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:87:30 (38)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  4:
    0: clearRect(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:90:26 (40)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  5:
    0: clearShadow()
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:93:28 (42)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  6:
    0: clip("nonzero")
      swizzleTypes: [String (44)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:96:21 (45)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: clip("evenodd")
      swizzleTypes: [String (46)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:97:21 (47)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: clip([object Path2D], "nonzero")
      swizzleTypes: [Path2D (48), String (44)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:98:21 (49)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    3: clip([object Path2D], "evenodd")
      swizzleTypes: [Path2D (50), String (46)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:99:21 (51)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  7:
    0: closePath()
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:102:26 (53)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  8:
    0: createImageData([object ImageData])
      swizzleTypes: [ImageData (55)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:105:54 (56)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:105:28 (57)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: createImageData(2, 3)
      swizzleTypes: [Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:106:54 (58)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:106:28 (59)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  9:
    0: createLinearGradient(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:109:59 (61)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:109:28 (62)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  10:
    0: createPattern([object HTMLImageElement], "testA")
      swizzleTypes: [Image (64), String (65)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:112:52 (66)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:112:28 (67)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: createPattern([object HTMLImageElement], "testB")
      swizzleTypes: [Image (68), String (69)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:113:52 (70)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:113:28 (71)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: createPattern([object HTMLImageElement], "testC")
      swizzleTypes: [Image (72), String (73)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:114:52 (74)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:114:28 (75)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  11:
    0: createRadialGradient(1, 2, 3, 4, 5, 6)
      swizzleTypes: [Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:117:59 (77)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:117:28 (78)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  12:
    0: direction
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:120:16 (80)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  13:
    0: drawFocusIfNeeded("Element")
      swizzleTypes: [None (82)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:124:34 (83)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: drawFocusIfNeeded([object Path2D], "Element")
      swizzleTypes: [Path2D (48), None (82)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:125:34 (84)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  14:
    0: drawImage([object HTMLImageElement], 1, 2)
      swizzleTypes: [Image (64), Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:128:48 (86)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:128:28 (87)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: drawImage([object HTMLImageElement], 3, 4)
      swizzleTypes: [Image (68), Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:129:48 (88)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:129:28 (89)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: drawImage([object HTMLImageElement], 5, 6)
      swizzleTypes: [Image (72), Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:130:48 (90)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:130:28 (91)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    3: drawImage([object HTMLImageElement], 7, 8, 9, 10)
      swizzleTypes: [Image (64), Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:131:48 (92)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:131:28 (93)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    4: drawImage([object HTMLImageElement], 11, 12, 13, 14)
      swizzleTypes: [Image (68), Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:132:48 (94)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:132:28 (95)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    5: drawImage([object HTMLImageElement], 15, 16, 17, 18)
      swizzleTypes: [Image (72), Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:133:48 (96)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:133:28 (97)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    6: drawImage([object HTMLImageElement], 19, 20, 21, 22, 23, 24, 25, 26)
      swizzleTypes: [Image (64), Number, Number, Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:134:48 (98)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:134:28 (99)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    7: drawImage([object HTMLImageElement], 27, 28, 29, 30, 31, 32, 33, 34)
      swizzleTypes: [Image (68), Number, Number, Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:135:48 (100)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:135:28 (101)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    8: drawImage([object HTMLImageElement], 35, 36, 37, 38, 39, 40, 41, 42)
      swizzleTypes: [Image (72), Number, Number, Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:136:48 (102)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:136:28 (103)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  15:
    0: drawImageFromRect([object HTMLImageElement], 1, 2, 3, 4, 5, 6, 7, 8, "")
      swizzleTypes: [Image (64), Number, Number, Number, Number, Number, Number, Number, Number, String (10)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:139:34 (105)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: drawImageFromRect([object HTMLImageElement], 9, 10, 11, 12, 13, 14, 15, 16, "test")
      swizzleTypes: [Image (64), Number, Number, Number, Number, Number, Number, Number, Number, String (106)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:140:34 (107)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  16:
    0: ellipse(1, 2, 3, 4, 5, 6, 7, false)
      swizzleTypes: [Number, Number, Number, Number, Number, Number, Number, Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:143:46 (109)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:143:28 (110)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: ellipse(8, 9, 10, 11, 12, 13, 14, true)
      swizzleTypes: [Number, Number, Number, Number, Number, Number, Number, Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:144:46 (111)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:144:28 (112)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  17:
    0: fill("nonzero")
      swizzleTypes: [String (44)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:147:21 (114)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: fill("evenodd")
      swizzleTypes: [String (46)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:148:21 (115)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: fill([object Path2D], "nonzero")
      swizzleTypes: [Path2D (48), String (44)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:149:21 (116)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    3: fill([object Path2D], "evenodd")
      swizzleTypes: [Path2D (50), String (46)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:150:21 (117)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  18:
    0: fillRect(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:153:25 (119)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  19:
    0: fillStyle
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:156:16 (121)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: fillStyle = "test"
      swizzleTypes: [String (106)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:157:16 (122)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: fillStyle = [object CanvasGradient]
      swizzleTypes: [CanvasGradient (124)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:158:16 (125)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    3: fillStyle = [object CanvasGradient]
      swizzleTypes: [CanvasGradient (127)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:159:16 (128)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    4: fillStyle = [object CanvasPattern]
      swizzleTypes: [CanvasPattern (131)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:160:16 (132)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  20:
    0: fillText("testA", 1, 2)
      swizzleTypes: [String (65), Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:163:25 (134)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: fillText("testB", 3, 4, 5)
      swizzleTypes: [String (69), Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:164:25 (135)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  21:
    0: font
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:167:16 (137)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: font = "test"
      swizzleTypes: [String (106)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:168:16 (138)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  22:
    0: getImageData(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:171:51 (140)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:171:28 (141)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  23:
    0: getLineDash()
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:174:28 (143)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  24:
    0: getTransform
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:177:29 (145)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  25:
    0: globalAlpha
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:180:16 (147)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: globalAlpha = 0
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:181:16 (148)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  26:
    0: globalCompositeOperation
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:184:16 (150)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: globalCompositeOperation = "test"
      swizzleTypes: [String (106)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:185:16 (151)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  27:
    0: imageSmoothingEnabled
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:188:16 (153)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: imageSmoothingEnabled = true
      swizzleTypes: [Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:189:16 (154)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  28:
    0: imageSmoothingQuality
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:192:16 (156)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: imageSmoothingQuality = "low"
      swizzleTypes: [String (9)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:193:16 (157)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  29:
    0: isPointInPath([object Path2D], 5, 6, "nonzero")
      swizzleTypes: [Path2D (48), Number, Number, String (44)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:196:30 (159)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: isPointInPath([object Path2D], 7, 8, "evenodd")
      swizzleTypes: [Path2D (50), Number, Number, String (46)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:197:30 (160)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: isPointInPath(9, 10, "nonzero")
      swizzleTypes: [Number, Number, String (44)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:198:30 (161)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    3: isPointInPath(11, 12, "evenodd")
      swizzleTypes: [Number, Number, String (46)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:199:30 (162)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  30:
    0: isPointInStroke([object Path2D], 3, 4)
      swizzleTypes: [Path2D (48), Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:202:32 (164)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: isPointInStroke(5, 6)
      swizzleTypes: [Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:203:32 (165)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  31:
    0: lineCap
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:206:16 (167)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  32:
    0: lineDashOffset
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:210:16 (169)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: lineDashOffset = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:211:16 (170)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  33:
    0: lineJoin
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:214:16 (172)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  34:
    0: lineTo(1, 2)
      swizzleTypes: [Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:218:23 (174)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  35:
    0: lineWidth
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:221:16 (176)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: lineWidth = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:222:16 (177)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  36:
    0: measureText("test")
      swizzleTypes: [String (106)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:225:28 (179)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  37:
    0: miterLimit
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:228:16 (181)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: miterLimit = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:229:16 (182)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  38:
    0: moveTo(1, 2)
      swizzleTypes: [Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:232:23 (184)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  39:
    0: putImageData([object ImageData], 5, 6)
      swizzleTypes: [ImageData (55), Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:235:29 (186)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: putImageData([object ImageData], 7, 8, 9, 10, 11, 12)
      swizzleTypes: [ImageData (187), Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:236:29 (188)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  40:
    0: quadraticCurveTo(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:239:33 (190)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  41:
    0: rect(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:242:21 (192)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  42:
    0: resetTransform()
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:245:31 (194)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  43:
    0: restore()
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:248:24 (196)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  44:
    0: rotate(1)
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:253:23 (198)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  45:
    0: save()
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:256:21 (200)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  46:
    0: scale(1, 2)
      swizzleTypes: [Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:259:22 (202)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  47:
    0: setAlpha(null)
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:262:25 (204)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: setAlpha(1)
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:263:25 (205)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  48:
    0: setCompositeOperation("undefined")
      swizzleTypes: [String (207)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:266:38 (208)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: setCompositeOperation("test")
      swizzleTypes: [String (106)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:267:38 (209)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  49:
    0: setFillColor("testA")
      swizzleTypes: [String (65)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:270:29 (211)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: setFillColor("testB", 1)
      swizzleTypes: [String (69), Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:271:29 (212)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: setFillColor(2, 1)
      swizzleTypes: [Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:272:29 (213)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    3: setFillColor(3, 4)
      swizzleTypes: [Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:273:29 (214)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    4: setFillColor(5, 6, 7, 8)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:274:29 (215)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    5: setFillColor(9, 10, 11, 12, 13)
      swizzleTypes: [Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:275:29 (216)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  50:
    0: setLineCap("undefined")
      swizzleTypes: [String (207)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:278:27 (218)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: setLineCap("test")
      swizzleTypes: [String (106)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:279:27 (219)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  51:
    0: setLineDash([1,2])
      swizzleTypes: [Array]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:282:28 (221)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  52:
    0: setLineJoin("undefined")
      swizzleTypes: [String (207)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:285:28 (223)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: setLineJoin("test")
      swizzleTypes: [String (106)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:286:28 (224)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  53:
    0: setLineWidth(null)
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:289:29 (226)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: setLineWidth(1)
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:290:29 (227)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  54:
    0: setMiterLimit(null)
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:293:30 (229)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: setMiterLimit(1)
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:294:30 (230)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  55:
    0: setShadow(1, 2, 3, "")
      swizzleTypes: [Number, Number, Number, String (232)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:297:26 (233)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: setShadow(4, 5, 6, "test", 7)
      swizzleTypes: [Number, Number, Number, String (106), Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:298:26 (234)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: setShadow(8, 9, 10, 11, 1)
      swizzleTypes: [Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:299:26 (235)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    3: setShadow(12, 13, 14, 15, 16)
      swizzleTypes: [Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:300:26 (236)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    4: setShadow(17, 18, 19, 20, 21, 22, 23)
      swizzleTypes: [Number, Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:301:26 (237)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    5: setShadow(24, 25, 26, 27, 28, 29, 30, 31)
      swizzleTypes: [Number, Number, Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:302:26 (238)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  56:
    0: setStrokeColor("testA")
      swizzleTypes: [String (65)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:305:31 (240)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: setStrokeColor("testB", 1)
      swizzleTypes: [String (69), Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:306:31 (241)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: setStrokeColor(2, 1)
      swizzleTypes: [Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:307:31 (242)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    3: setStrokeColor(3, 4)
      swizzleTypes: [Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:308:31 (243)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    4: setStrokeColor(5, 6, 7, 8)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:309:31 (244)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    5: setStrokeColor(9, 10, 11, 12, 13)
      swizzleTypes: [Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:310:31 (245)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  57:
    0: setTransform(1, 2, 3, 4, 5, 6)
      swizzleTypes: [Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:313:29 (247)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: setTransform(matrix(1, 0, 0, 1, 0, 0))
      swizzleTypes: [DOMMatrix]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:314:51 (248)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:314:28 (249)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: setTransform(matrix(7, 8, 9, 10, 11, 12))
      swizzleTypes: [DOMMatrix]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:315:51 (250)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:315:28 (251)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  58:
    0: shadowBlur
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:318:16 (253)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: shadowBlur = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:319:16 (254)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  59:
    0: shadowColor
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:322:16 (256)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: shadowColor = "test"
      swizzleTypes: [String (106)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:323:16 (257)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  60:
    0: shadowOffsetX
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:326:16 (259)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: shadowOffsetX = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:327:16 (260)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  61:
    0: shadowOffsetY
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:330:16 (262)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: shadowOffsetY = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:331:16 (263)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  62:
    0: stroke()
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:334:23 (265)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: stroke([object Path2D])
      swizzleTypes: [Path2D (48)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:335:23 (266)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  63:
    0: strokeRect(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:338:27 (268)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  64:
    0: strokeStyle
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:341:16 (270)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: strokeStyle = "test"
      swizzleTypes: [String (106)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:342:16 (271)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    2: strokeStyle = [object CanvasGradient]
      swizzleTypes: [CanvasGradient (124)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:343:16 (272)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    3: strokeStyle = [object CanvasGradient]
      swizzleTypes: [CanvasGradient (127)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:344:16 (273)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    4: strokeStyle = [object CanvasPattern]
      swizzleTypes: [CanvasPattern (131)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:345:16 (274)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  65:
    0: strokeText("testA", 1, 2)
      swizzleTypes: [String (65), Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:348:27 (276)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: strokeText("testB", 3, 4, 5)
      swizzleTypes: [String (69), Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:349:27 (277)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  66:
    0: textAlign
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:352:16 (279)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  67:
    0: textBaseline
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:356:16 (281)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  68:
    0: transform(1, 2, 3, 4, 5, 6)
      swizzleTypes: [Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:360:26 (283)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  69:
    0: translate(1, 2)
      swizzleTypes: [Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:363:26 (285)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  70:
    0: webkitGetImageDataHD(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:366:59 (287)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:366:28 (288)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  71:
    0: webkitImageSmoothingEnabled
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:369:16 (290)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: webkitImageSmoothingEnabled = true
      swizzleTypes: [Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:370:16 (291)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  72:
    0: webkitLineDash
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:373:16 (293)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: webkitLineDash = [1,2]
      swizzleTypes: [Array]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:374:16 (294)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  73:
    0: webkitLineDashOffset
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:377:16 (296)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: webkitLineDashOffset = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:378:16 (297)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
  74:
    0: webkitPutImageDataHD([object ImageData], 5, 6)
      swizzleTypes: [ImageData (55), Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:381:37 (299)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
    1: webkitPutImageDataHD([object ImageData], 7, 8, 9, 10, 11, 12)
      swizzleTypes: [ImageData (187), Number, Number, Number, Number, Number, Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:382:37 (300)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
data:
  0: "source-over"
  1: "butt"
  2: "miter"
  3: "rgba(0, 0, 0, 0)"
  4: "10px sans-serif"
  5: "start"
  6: "alphabetic"
  7: "ltr"
  8: "#000000"
  9: "low"
  10: ""
  11: "arc"
  12: "inspector/canvas/recording-2d.html"
  13: [10,12,78,42]
  14: "ignoreException"
  15: [14,12,49,13]
  16: [10,12,78,28]
  17: "executeFrameFunction"
  18: [17,12,391,24]
  19: "performActions"
  20: [19,12,395,25]
  21: "global code"
  22: [21,10,1,15]
  23: "evaluateWithScopeExtension"
  24: "[native code]"
  25: [23,24,0,0]
  26: "_evaluateOn"
  27: [26,10,128,29]
  28: "_evaluateAndWrap"
  29: [28,10,122,108]
  30: [10,12,79,42]
  31: [10,12,79,28]
  32: "arcTo"
  33: [10,12,82,44]
  34: [10,12,82,28]
  35: "beginPath"
  36: [10,12,85,26]
  37: "bezierCurveTo"
  38: [10,12,88,30]
  39: "clearRect"
  40: [10,12,91,26]
  41: "clearShadow"
  42: [10,12,94,28]
  43: "clip"
  44: "nonzero"
  45: [10,12,97,21]
  46: "evenodd"
  47: [10,12,98,21]
  48: "M1 2"
  49: [10,12,99,21]
  50: "M3 4"
  51: [10,12,100,21]
  52: "closePath"
  53: [10,12,103,26]
  54: "createImageData"
  55: [[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],1,4]
  56: [10,12,106,54]
  57: [10,12,106,28]
  58: [10,12,107,54]
  59: [10,12,107,28]
  60: "createLinearGradient"
  61: [10,12,110,59]
  62: [10,12,110,28]
  63: "createPattern"
  64: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAABNJREFUCB1j/M/AAEQMDEwgAgQAHxcCAmtAm/sAAAAASUVORK5CYII="
  65: "testA"
  66: [10,12,113,52]
  67: [10,12,113,28]
  68: "data:,"
  69: "testB"
  70: [10,12,114,52]
  71: [10,12,114,28]
  72: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAAAtJREFUCB1jYEAHAAASAAGAFMrMAAAAAElFTkSuQmCC"
  73: "testC"
  74: [10,12,115,52]
  75: [10,12,115,28]
  76: "createRadialGradient"
  77: [10,12,118,59]
  78: [10,12,118,28]
  79: "direction"
  80: [10,12,121,16]
  81: "drawFocusIfNeeded"
  82: "Element"
  83: [10,12,125,34]
  84: [10,12,126,34]
  85: "drawImage"
  86: [10,12,129,48]
  87: [10,12,129,28]
  88: [10,12,130,48]
  89: [10,12,130,28]
  90: [10,12,131,48]
  91: [10,12,131,28]
  92: [10,12,132,48]
  93: [10,12,132,28]
  94: [10,12,133,48]
  95: [10,12,133,28]
  96: [10,12,134,48]
  97: [10,12,134,28]
  98: [10,12,135,48]
  99: [10,12,135,28]
  100: [10,12,136,48]
  101: [10,12,136,28]
  102: [10,12,137,48]
  103: [10,12,137,28]
  104: "drawImageFromRect"
  105: [10,12,140,34]
  106: "test"
  107: [10,12,141,34]
  108: "ellipse"
  109: [10,12,144,46]
  110: [10,12,144,28]
  111: [10,12,145,46]
  112: [10,12,145,28]
  113: "fill"
  114: [10,12,148,21]
  115: [10,12,149,21]
  116: [10,12,150,21]
  117: [10,12,151,21]
  118: "fillRect"
  119: [10,12,154,25]
  120: "fillStyle"
  121: [10,12,157,16]
  122: [10,12,158,16]
  123: "linear-gradient"
  124: [123,[1,2,3,4],[]]
  125: [10,12,159,16]
  126: "radial-gradient"
  127: [126,[1,2,3,4,5,6],[]]
  128: [10,12,160,16]
  129: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAABNJREFUCB1j/M/AAEQMDEwgAgQAHxcCAmtAm/sAAAAASUVORK5CYII="
  130: "no-repeat"
  131: [129,130]
  132: [10,12,161,16]
  133: "fillText"
  134: [10,12,164,25]
  135: [10,12,165,25]
  136: "font"
  137: [10,12,168,16]
  138: [10,12,169,16]
  139: "getImageData"
  140: [10,12,172,51]
  141: [10,12,172,28]
  142: "getLineDash"
  143: [10,12,175,28]
  144: "getTransform"
  145: [10,12,178,29]
  146: "globalAlpha"
  147: [10,12,181,16]
  148: [10,12,182,16]
  149: "globalCompositeOperation"
  150: [10,12,185,16]
  151: [10,12,186,16]
  152: "imageSmoothingEnabled"
  153: [10,12,189,16]
  154: [10,12,190,16]
  155: "imageSmoothingQuality"
  156: [10,12,193,16]
  157: [10,12,194,16]
  158: "isPointInPath"
  159: [10,12,197,30]
  160: [10,12,198,30]
  161: [10,12,199,30]
  162: [10,12,200,30]
  163: "isPointInStroke"
  164: [10,12,203,32]
  165: [10,12,204,32]
  166: "lineCap"
  167: [10,12,207,16]
  168: "lineDashOffset"
  169: [10,12,211,16]
  170: [10,12,212,16]
  171: "lineJoin"
  172: [10,12,215,16]
  173: "lineTo"
  174: [10,12,219,23]
  175: "lineWidth"
  176: [10,12,222,16]
  177: [10,12,223,16]
  178: "measureText"
  179: [10,12,226,28]
  180: "miterLimit"
  181: [10,12,229,16]
  182: [10,12,230,16]
  183: "moveTo"
  184: [10,12,233,23]
  185: "putImageData"
  186: [10,12,236,29]
  187: [[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],2,3]
  188: [10,12,237,29]
  189: "quadraticCurveTo"
  190: [10,12,240,33]
  191: "rect"
  192: [10,12,243,21]
  193: "resetTransform"
  194: [10,12,246,31]
  195: "restore"
  196: [10,12,249,24]
  197: "rotate"
  198: [10,12,254,23]
  199: "save"
  200: [10,12,257,21]
  201: "scale"
  202: [10,12,260,22]
  203: "setAlpha"
  204: [10,12,263,25]
  205: [10,12,264,25]
  206: "setCompositeOperation"
  207: "undefined"
  208: [10,12,267,38]
  209: [10,12,268,38]
  210: "setFillColor"
  211: [10,12,271,29]
  212: [10,12,272,29]
  213: [10,12,273,29]
  214: [10,12,274,29]
  215: [10,12,275,29]
  216: [10,12,276,29]
  217: "setLineCap"
  218: [10,12,279,27]
  219: [10,12,280,27]
  220: "setLineDash"
  221: [10,12,283,28]
  222: "setLineJoin"
  223: [10,12,286,28]
  224: [10,12,287,28]
  225: "setLineWidth"
  226: [10,12,290,29]
  227: [10,12,291,29]
  228: "setMiterLimit"
  229: [10,12,294,30]
  230: [10,12,295,30]
  231: "setShadow"
  232: ""
  233: [10,12,298,26]
  234: [10,12,299,26]
  235: [10,12,300,26]
  236: [10,12,301,26]
  237: [10,12,302,26]
  238: [10,12,303,26]
  239: "setStrokeColor"
  240: [10,12,306,31]
  241: [10,12,307,31]
  242: [10,12,308,31]
  243: [10,12,309,31]
  244: [10,12,310,31]
  245: [10,12,311,31]
  246: "setTransform"
  247: [10,12,314,29]
  248: [10,12,315,51]
  249: [10,12,315,28]
  250: [10,12,316,51]
  251: [10,12,316,28]
  252: "shadowBlur"
  253: [10,12,319,16]
  254: [10,12,320,16]
  255: "shadowColor"
  256: [10,12,323,16]
  257: [10,12,324,16]
  258: "shadowOffsetX"
  259: [10,12,327,16]
  260: [10,12,328,16]
  261: "shadowOffsetY"
  262: [10,12,331,16]
  263: [10,12,332,16]
  264: "stroke"
  265: [10,12,335,23]
  266: [10,12,336,23]
  267: "strokeRect"
  268: [10,12,339,27]
  269: "strokeStyle"
  270: [10,12,342,16]
  271: [10,12,343,16]
  272: [10,12,344,16]
  273: [10,12,345,16]
  274: [10,12,346,16]
  275: "strokeText"
  276: [10,12,349,27]
  277: [10,12,350,27]
  278: "textAlign"
  279: [10,12,353,16]
  280: "textBaseline"
  281: [10,12,357,16]
  282: "transform"
  283: [10,12,361,26]
  284: "translate"
  285: [10,12,364,26]
  286: "webkitGetImageDataHD"
  287: [10,12,367,59]
  288: [10,12,367,28]
  289: "webkitImageSmoothingEnabled"
  290: [10,12,370,16]
  291: [10,12,371,16]
  292: "webkitLineDash"
  293: [10,12,374,16]
  294: [10,12,375,16]
  295: "webkitLineDashOffset"
  296: [10,12,378,16]
  297: [10,12,379,16]
  298: "webkitPutImageDataHD"
  299: [10,12,382,37]
  300: [10,12,383,37]

-- Running test case: Canvas.recording2D.memoryLimit
initialState:
  attributes:
    width: 2
    height: 2
    setTransform: [1,0,0,1,0,0]
    globalAlpha: 1
    globalCompositeOperation: 0
    lineWidth: 1
    lineCap: 1
    lineJoin: 2
    miterLimit: 10
    shadowOffsetX: 0
    shadowOffsetY: 0
    shadowBlur: 0
    shadowColor: 3
    setLineDash: [[]]
    lineDashOffset: 0
    font: 4
    textAlign: 5
    textBaseline: 6
    direction: 7
    strokeStyle: 8
    fillStyle: 8
    imageSmoothingEnabled: true
    imageSmoothingQuality: 9
    setPath: [10]
  parameters:
  content: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAAAtJREFUCB1jYEAHAAASAAGAFMrMAAAAAElFTkSuQmCC"
frames:
  0: (incomplete)
    0: arc(1, 2, 3, 4, 5, false)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:77:42 (13)
        1: ignoreException - inspector/canvas/recording-2d.html:48:13 (15)
        2: (anonymous function) - inspector/canvas/recording-2d.html:77:28 (16)
        3: executeFrameFunction - inspector/canvas/recording-2d.html:390:24 (18)
        4: performActions - inspector/canvas/recording-2d.html:394:25 (20)
        5: Global Code - [program code] (22)
        6: evaluateWithScopeExtension - [native code] (25)
        7: _evaluateOn (27)
        8: _evaluateAndWrap (29)
data:
  0: "source-over"
  1: "butt"
  2: "miter"
  3: "rgba(0, 0, 0, 0)"
  4: "10px sans-serif"
  5: "start"
  6: "alphabetic"
  7: "ltr"
  8: "#000000"
  9: "low"
  10: ""
  11: "arc"
  12: "inspector/canvas/recording-2d.html"
  13: [10,12,78,42]
  14: "ignoreException"
  15: [14,12,49,13]
  16: [10,12,78,28]
  17: "executeFrameFunction"
  18: [17,12,391,24]
  19: "performActions"
  20: [19,12,395,25]
  21: "global code"
  22: [21,10,1,15]
  23: "evaluateWithScopeExtension"
  24: "[native code]"
  25: [23,24,0,0]
  26: "_evaluateOn"
  27: [26,10,128,29]
  28: "_evaluateAndWrap"
  29: [28,10,122,108]

-- Running test case: Canvas.recording2D.ActionParameterNaN
PASS: The recording should have 1 frame.
PASS: The first frame should have 1 action.
PASS: The action should have 1 parameter.
PASS: The parameter should be null.
